Refactor test helpers into their own class.
