+++ src/Band.Personalize/Band.Personalize.Model.Test/Band.Personalize.Model.Test.csproj
@@ -128,6 +128,7 @@
     <Compile Include="Repository\BandRepositoryTests.cs" />
     <Compile Include="Repository\BandTests.cs" />
     <Compile Include="Repository\IBandClientManagerExtensionsTests.cs" />
+    <Compile Include="TestHelper.cs" />
     <Compile Include="UnitTestApp.xaml.cs">
       <DependentUpon>UnitTestApp.xaml</DependentUpon>
     </Compile>
